Add MSF history page
